re-arrange some codes
